+++ e2e/testing/ibc_path.go
@@ -335,9 +335,7 @@ func (e *IBCEndpoint) sendConnectionOpenInit() {
 	const (
 		defDelayPeriod uint64 = 0
 	)
-	var (
-		version = connectionTypes.ExportedVersionsToProto(connectionTypes.GetCompatibleVersions())[0]
-	)
+	version := connectionTypes.ExportedVersionsToProto(connectionTypes.GetCompatibleVersions())[0]
 
 	t, srcChain, dstChain := e.t, e.srcChain, e.dstEndpoint.srcChain
 	srcChainClientID, dstChainClientID := e.clientID, e.dstEndpoint.clientID
@@ -364,9 +362,7 @@ func (e *IBCEndpoint) sendConnectionOpenTry() {
 	const (
 		defDelayPeriod uint64 = 0
 	)
-	var (
-		version = connectionTypes.ExportedVersionsToProto(connectionTypes.GetCompatibleVersions())[0]
-	)
+	version := connectionTypes.ExportedVersionsToProto(connectionTypes.GetCompatibleVersions())[0]
 
 	t, srcChain, dstChain := e.t, e.srcChain, e.dstEndpoint.srcChain
 	srcChainClientID, dstChainClientID, srcChainConnectionID, dstChainConnectionID := e.clientID, e.dstEndpoint.clientID, e.connectionID, e.dstEndpoint.connectionID
@@ -404,9 +400,7 @@ func (e *IBCEndpoint) sendConnectionOpenTry() {
 
 // sendConnectionOpenAck sends a ConnectionOpenAck message to the source chain.
 func (e *IBCEndpoint) sendConnectionOpenAck() {
-	var (
-		version = connectionTypes.ExportedVersionsToProto(connectionTypes.GetCompatibleVersions())[0]
-	)
+	version := connectionTypes.ExportedVersionsToProto(connectionTypes.GetCompatibleVersions())[0]
 
 	srcChain := e.srcChain
 	srcChainConnectionID, dstChainConnectionID := e.connectionID, e.dstEndpoint.connectionID