+++ x/rewards/keeper/flat_fee.go
@@ -41,3 +41,14 @@ func (k Keeper) GetFlatFee(ctx sdk.Context, contractAddr sdk.AccAddress) (sdk.Co
 
 	return fee, true
 }
+
+// CreateFlatFeeRewardsRecords creates a rewards record for the flatfees of the given contract
+func (k Keeper) CreateFlatFeeRewardsRecords(ctx sdk.Context, contractAddress sdk.AccAddress, flatfees sdk.Coins) {
+	rewardsRecordState := k.state.RewardsRecord(ctx)
+	calculationHeight, calculationTime := ctx.BlockHeight(), ctx.BlockTime()
+
+	metadata := k.GetContractMetadata(ctx, contractAddress)
+	rewardsAddr := sdk.MustAccAddressFromBech32(metadata.RewardsAddress)
+
+	rewardsRecordState.CreateRewardsRecord(rewardsAddr, flatfees, calculationHeight, calculationTime)
+}